requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.1
colorama==0.4.4
cryptography==36.0.2
cssselect==1.1.0
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
geckodriver-autoinstaller==0.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
phantomjs==1.4.1
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
selenium==4.1.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.2.13
tinysegmenter==0.3
tldextract==3.2.0
tqdm==4.63.1
trio==0.20.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.9
webdriver-manager==3.5.4
webencodings==0.5.1
websockets==10.2
wsproto==1.1.0
xmltodict==0.12.0
xvfbwrapper==0.2.9
